application/front/requirements.txt
streamlit==1.39.0
requests==2.32.3
pillow==10.4.0